% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory_systems.R
\name{create_directory_system}
\alias{create_directory_system}
\title{Create the appropriate directory system to be able to run functions without specifying hard filepaths}
\usage{
create_directory_system(rootdir = NULL)
}
\arguments{
\item{rootdir}{Directory within which to create the directory system}
}
\value{
No return value, creates directory system in the specified directory.
}
\description{
Create the appropriate directory system to be able to run functions without specifying hard filepaths
}
\examples{
## Create directory system compatible with rcprd's automatic saving of output
create_directory_system(tempdir())
file.exists(file.path(tempdir(),"data"))
file.exists(file.path(tempdir(),"code"))
file.exists(file.path(tempdir(),"codelists"))

## Return filespace to how it was prior to example
delete_directory_system(tempdir())

}
