% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{extract_practices}
\alias{extract_practices}
\title{Combine practice files}
\usage{
extract_practices(filepath, select = NULL)
}
\arguments{
\item{filepath}{Path to directory containing practice .txt files.}

\item{select}{Character vector of column names to select.}
}
\value{
Data frame with patient information
}
\description{
Combine practice files
}
\examples{

## Extract cohort data
prac<-extract_practices(filepath = system.file("aurum_data", package = "rcprd"))
prac

}
