% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{implement_output}
\alias{implement_output}
\title{Internal function to implement saving extracted variable to disk or returning into R workspace.}
\usage{
implement_output(
  variable_dat,
  varname,
  out_save_disk,
  out_subdir,
  out_filepath,
  return_output
)
}
\arguments{
\item{variable_dat}{Dataset containing variable}

\item{varname}{Name of variable to use in filename}

\item{out_save_disk}{If TRUE will save output to disk}

\item{out_subdir}{Sub-directory of data/ to save output into}

\item{out_filepath}{Full fiilepath to save dat onto}

\item{return_output}{If TRUE returns output into R workspace}
}
\description{
Will save extracted variable to disk if \code{out_save_disk = TRUE}. Note it relies on correct underlying structure
of directories. Will output extracted variable into R workspace if \code{return_output = TRUE}.
}
