% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_read_txt.R
\name{extract_txt_prac}
\alias{extract_txt_prac}
\title{Read in raw .txt practice file}
\usage{
extract_txt_prac(filepath, ..., select = NULL)
}
\arguments{
\item{filepath}{File path to raw .txt file}

\item{...}{Arguments to pass onto data.table::fread}

\item{select}{Character vector of variable names to select}
}
\description{
Read in raw .txt practice file
}
